♻️ refactor :  CloseButton
