+++ src/app/(Sub)/menu/page.tsx
@@ -1,7 +1,7 @@
 import LoginSection from './components/LoginSection';
 import MakerSection from './components/MakerSection';
-import MenuHeader from './components/MenuListSection';
-import MenuListSection from './components/MenuSection';
+import MenuHeader from './components/MenuHeader';
+import MenuListSection from './components/MenuListSection';
 
 export default function MenuPage() {
   return (
